refactor(app): move buildroot update to top-level substate
