+++ app/src/buildroot/__tests__/epic.test.js
@@ -4,15 +4,15 @@ import {
   makeRobotApiRequest,
   passRobotApiResponseAction,
   passRobotApiErrorAction,
-} from '../../../robot-api/deprecated'
+} from '../../robot-api/deprecated'
 
-import * as epics from '../update-epics'
+import * as epics from '../epic'
 import * as actions from '../actions'
 import * as selectors from '../selectors'
 import { INITIAL_STATE } from '../reducer'
 
 jest.mock('../selectors')
-jest.mock('../../../robot-api/deprecated')
+jest.mock('../../robot-api/deprecated')
 
 const robot = { name: 'robot', host: '10.10.0.0', port: 31950 }
