+++ app/src/components/RobotSettings/UpdateBuildroot/InstallModal.js
@@ -5,7 +5,10 @@ import { AlertModal } from '@opentrons/components'
 import InstallModalContents from './InstallModalContents'
 
 import type { ViewableRobot } from '../../../discovery/types'
-import type { BuildrootUpdateSession, RobotSystemType } from '../../../shell'
+import type {
+  BuildrootUpdateSession,
+  RobotSystemType,
+} from '../../../buildroot/types'
 
 type Props = {|
   robot: ViewableRobot,
