+++ app/src/components/RobotSettings/UpdateBuildroot/ViewUpdateModal.js
@@ -6,13 +6,16 @@ import {
   getBuildrootUpdateInfo,
   getBuildrootDownloadProgress,
   getBuildrootDownloadError,
-} from '../../../shell'
+} from '../../../buildroot'
 
 import MigrationWarningModal from './MigrationWarningModal'
 import DownloadUpdateModal from './DownloadUpdateModal'
 import ReleaseNotesModal from './ReleaseNotesModal'
 
-import type { BuildrootUpdateType, RobotSystemType } from '../../../shell'
+import type {
+  BuildrootUpdateType,
+  RobotSystemType,
+} from '../../../buildroot/types'
 
 type Props = {|
   robotName: string,
