+++ app/src/nav/selectors.js
@@ -4,7 +4,8 @@ import { createSelector } from 'reselect'
 import { getConnectedRobot } from '../discovery'
 import { getProtocolPipettesMatch } from '../pipettes'
 import { selectors as RobotSelectors } from '../robot'
-import { getAvailableShellUpdate, getBuildrootUpdateAvailable } from '../shell'
+import { getBuildrootUpdateAvailable } from '../buildroot'
+import { getAvailableShellUpdate } from '../shell'
 
 import type { State } from '../types'
 import type { NavLocation } from './types'
