--- app/src/shell/buildroot/index.js
@@ -1,7 +1,0 @@
-// @flow
-
-export * from './types'
-export * from './actions'
-export * from './reducer'
-export * from './selectors'
-export { buildrootUpdateEpic } from './update-epics'
